requirements.txt
PyQt5
PyOpenGL
vulkan
